requirements.txt
mysql-connector==2.2.9
mysql-connector-python==8.0.31
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
streamlit==1.14.0
streamlit-option-menu==0.3.2
